feature(Refresh button): Added a Refresh button on the Scaffold Top bar.
Refresh button clears local DB and re-downloads it

+++ lib/feature/home/data/model/date.dart
@@ -26,7 +26,7 @@ class Date {
       name: json['Name'] as String,
       category: Category.values.byName(json['Category']),
       suit: Suit.values.byName(json['Suit']),
-      effortValue: json['EffortValue'],
+      effortValue: json['EffortValue'] as int,
       favorite: json['Favorite'] as int,
     );
   }
